chore(storybook): integrate Tailwind CSS and restructure stories
