+++ apps/storybook/.storybook/preview.ts
@@ -1,5 +1,7 @@
 import type { Preview } from "@storybook/react-vite";
 
+import "./styles/global.css";
+
 const preview: Preview = {
   parameters: {
     controls: {
@@ -8,11 +10,7 @@ const preview: Preview = {
         date: /Date$/i,
       },
     },
-
     a11y: {
-      // 'todo' - show a11y violations in the test UI only
-      // 'error' - fail CI on a11y violations
-      // 'off' - skip a11y checks entirely
       test: "todo",
     },
   },
