+++ apps/storybook/.storybook/utils/get-absolute-path.ts
@@ -1,0 +1,7 @@
+import { join, dirname } from "path";
+
+function getAbsolutePath(value: string): any {
+  return dirname(require.resolve(join(value, "package.json")));
+}
+
+export default getAbsolutePath;
