+++ apps/storybook/.storybook/utils/index.ts
@@ -1,0 +1,3 @@
+import getAbsolutePath from "./get-absolute-path";
+
+export default getAbsolutePath;
