+++ apps/storybook/package.json
@@ -14,8 +14,11 @@
     "typescheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@tailwindcss/postcss": "^4.1.8",
+    "postcss": "^8.5.4",
     "react": "19.1.0",
-    "react-dom": "19.1.0"
+    "react-dom": "19.1.0",
+    "tailwindcss": "^4.1.8"
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^4.0.0",
