+++ apps/storybook/stories/button/button.stories.tsx
@@ -1,0 +1,26 @@
+import { fn } from "storybook/test";
+import type { Meta, StoryObj } from "@storybook/react-vite";
+
+import Button from "./button";
+
+const meta = {
+  title: "Example/Button",
+  component: Button,
+  parameters: {
+    layout: "centered",
+  },
+  tags: ["autodocs"],
+  argTypes: {
+    backgroundColor: { control: "color" },
+  },
+  args: { onClick: fn() },
+} satisfies Meta<typeof Button>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Primary: Story = {
+  args: {
+    label: "Button",
+  },
+};
